Refactor explainability
