+++ python_backend/models/lpdlprnet/api.py
@@ -1,7 +1,8 @@
 from app import app
 from flask import request, send_file, make_response
 from flask_cors import CORS, cross_origin
-from utils.utils import create_directories, check_request, crop_image, render_image
+from utils.utils import create_directories, check_request, crop_image, render_image, replace_in_markdown, save_image
+import pandas as pd
 
 import json
 import os
@@ -9,6 +10,7 @@ CORS(app)
 
 from models.lpdnet.lpd_model_class import LpdModelClass
 from models.lprnet.lpr_model_class import LprModelClass
+from models.lpdnet.api import evaluate_lpd
 
 @app.route('/api/lpdlprnet/<id>',methods= ['POST', 'GET'])
 def call_combined(id):
@@ -135,4 +137,114 @@ def call_combined(id):
                     processed[i]['overlay_image'] = f"{output_path}/overlay_lpdnet_{info['file_name']}"
                     
 
-        return make_response(processed,200)     
+        return make_response(processed,200)
+
+
+@app.route('/api/lpdlprnet/explain/<id>',methods= ['POST', 'GET'])
+def call_explain_combined(id):
+    """
+    This is a function to orchestrate the following steps: 
+    1. run lpd-lpr pipeline and spit out predictions
+    2. run evaluation of lpd and save overlay/results
+    3. replace placeholders in markdown file with their respective dynamic values
+    
+    """
+    #level of detail:
+    n = 11
+
+    lpd_mapping = json.load(open('/app/models/lpdlprnet/database/lpd_mapping.json'))
+    lpr_mapping = json.load(open('/app/models/lpdlprnet/database/lpr_mapping.json'))
+    LOGGING = True #Saves output images into the folders
+    LPD_THRESHOLD = 0.6 #Threshold to filter images
+    LPR_THRESHOLD = 0.6
+
+    if (id not in lpd_mapping) or (id not in lpr_mapping) or (id!='internal'): 
+        return make_response({'error':"Bad Request - Invalid ID"},400)
+
+    try:
+        lpr = LprModelClass(id, lpr_mapping[id])
+        lpd = LpdModelClass(id, lpd_mapping[id]) 
+    except ValueError:
+        return make_response({'error':"Model not found on triton server"},503)
+
+    # Load input images
+    files = request.files.to_dict(flat=False)['image']
+    
+    # Load filenames
+    filenames = request.form.getlist('filename')
+
+    output = check_request(request)
+
+    if output!=True: return make_response(output,400)
+    
+    # Create directories for input and output images
+    input_path, output_path = create_directories('lpdlprnet',id)
+
+    #STATIC FILENAMES
+    save_as = f"{output_path}/heatmap_{filenames[0]}"
+    baseimage = f"{input_path}/{filenames[0]}"
+    
+    images = {}
+
+    # Save input images --> for explain, only use 1st file
+    images[filenames[0]] = files[0]
+    files[0].save(baseimage)
+
+    # SEND TO LPD
+    lpd_response = lpd.predict(input_path)
+
+    processed = {}
+    reverse_mapping = {}
+    i, info = 0, lpd_response[0]
+
+    # DYNAMIC FILENAMES
+    lpdout = f"{output_path}/exp_{info['file_name']}"
+    demopic = f"{output_path}/overlay_lpdnet_{info['file_name']}"
+
+    if info['HTTPStatus']==204:
+        # No inference bounding box was found
+        processed[i] = info
+        save_image(images[info['file_name']], lpdout)
+        reverse_mapping[f"exp_{info['file_name']}"] = i
+
+    else:
+        # info is a list of bbox, bbox is a dict containing a list (bbox)
+        # and a single number, confidence score
+        for j, bbox_info in enumerate(info["all_bboxes"]):
+            crop_image(images[info['file_name']],bbox_info['bbox'],lpdout)
+            reverse_mapping[f"exp_{info['file_name']}"] = i
+            bbox_info[f"exp_bbox"] = bbox_info.pop('bbox')
+  
+
+    processed[i] = info
+
+    # Call LPR on output of LPD
+    lpr_response = lpr.predict(output_path)
+    
+    # Process response to return
+    for lpr_info in lpr_response:
+        file_name = lpr_info['file_name']
+        index = file_name.split("_")[0]
+        temp = {}
+        temp['license_plate'] = list(lpr_info['license_plate'])
+        temp['confidence_scores'] = lpr_info['confidence_scores']
+        processed[reverse_mapping[file_name]][f"exp_lpr"] = temp
+
+    # evaluate_lpd(baseimage, filenames[0], id, save_as, n)
+
+    if id=='internal': 
+        render_image(images[info['file_name']],info['all_bboxes'], demopic)
+        info['overlay_image'] = demopic   
+        
+    # replace markdown placeholders with custom images
+    image_replace = {
+        '%placeholder1%' : f"{baseimage}", 
+        '%placeholder2%' : f"{demopic}",
+        '%placeholder3%' : f"{lpdout}",
+        '%placeholder4%' : f"{save_as}",
+        '%placeholder5%' : pd.DataFrame(info['all_bboxes']).to_html(),
+        '%placeholder6%' : pd.DataFrame(temp).assign(lp = '')[['lp', 'license_plate', 'confidence_scores']].rename(columns={'lp':lpr_info['license_plate']}).to_html(index=False, ),
+    }
+    
+    
+    return make_response({'explain_markdown': replace_in_markdown(image_replace, "/app/models/lpdlprnet/database/lpdlprnet_explainability_dynamic.md")},200)
